+++ clox/clox-vm/src/clox/vm/dbg.h
@@ -6,6 +6,6 @@
 struct clox_vm_chunk;
 
 void clox_vm_dbg_chunk_disassemble(struct clox_vm_chunk* chunk, const char* name);
-int  clox_vm_dbg_chunk_disassemble_instruction(struct clox_vm_chunk* chunk, size_t offset);
+size_t clox_vm_dbg_chunk_disassemble_instruction(struct clox_vm_chunk* chunk, size_t offset);
 
 #endif
